fix: Store BTree entry size in the chunk  (#538)
## Problem

We need to know the size of an entry to know where to partition the nodes. We realized before that this is a performance bottle neck so we computed it once up front and cached it in memory. However, this means that we needed to compute the size of the entries when we read a btree chunk for writing. This was still too often and `getSizeOfValue` was still called to much.

## Solution

Store the size of the entry in the entry that we use in the the chunk store. Now an entry is `[key: string, value: V, sizeOfEntry: number]` inside the IDB.

### Mutation Recovery

This is a format change. We are now up to `REPLICACHE_FORMAT_VERSION_V7`. To allow recovering mutations of `V6` we need to be able to read and write both `V6` and `V7` BTree node chunks. To safely do this we pipe through the replicache format version through almost all layers of the system. When we read a `V6` btree node chunk entry we compute the size of the entry. For `V7` we just read it from the chunk. When writing a `V6` btree node chunk entry we strip the size.