feat!: Support fined-grained authorization logic on the server. (#589)
Enables fined-grained authorization logic on the server.

This is done by exposing the `UserData` for the authenticated connection over which a mutation was pushed to the mutator when it is run on the server.  This is exposed via a new userData property on new reflect specific ReadTransaction and WriteTransaction interfaces (note this property was not added to the Replicache ReadTransaction/WriteTransaction interfaces, as it is relevant to Replicache).  

`UserData` size is currently limited to 6kb.

#492 

BREAKING CHANGE: `UserData` type renamed to `AuthData` and its `userID` field is now readonly.