fix: Types for mirror-server (#611)
Problem
---

The problem is that some type definitions for `jest` are implicit which
requires adding `@types/jest` and some are explicit. But the types in
the jest libraries are not compatible with the types in the
`@types/jest` package (when using `exactOptionalPropertyTypes`).

`firestore-jest-mock` is a package that uses implicit dependencies so it
needs `@types/jest` and causes problems with and without `"types":
["jest"]` in `tsconfig.json`.

Solution
---

Use a `postinstall` npm script to replace the implicit jest dependencies
with a explicit one.