fix(reflect-server): msg context key breaks datadog logs (#632)
Problem
======
Confusingly, when Monday went from 0.25.1 to version 0.27.1  many log messages appeared to be being replaced with short random ids.

<img width="1297" alt="image" src="https://github.com/rocicorp/mono/assets/19158916/3c97a62a-c0ca-4e8f-9eaf-adcd4539eb86">

This was due to the change to send log context as attributes to datadog, in combination of the use of the string 'msg' as a context key.  

DataDog replaced the log message with the value of the 'msg' attribute. Evidently this is an undocumented reserved attribute.

Solution
======
Replace 'msg' context key with 'msgID'.  

Prefix reserved datadog attributes with '@DATADOG_RESERVED_' to avoid future confusion.