feat(reflect-server): improve logging of DO fetch errors (#698)
Problem
======
It is difficult to identify which layer of DO certain errors are coming from.   For example the following error log message with `component=AuthDO`:

`[Unhandled exception in fetch, {stack=Error: Durable Object is overloaded. Requests queued for too long., name=Error, message=Durable Object is overloaded. Requests queued for too long.}]`

Is it the AuthDO that is overloaded, or the RoomDO, if it is the RoomDO, what is the RoomDO's roomID?

Solution
======
Wrap fetches to DO's with error logging.