feat(reflect-client): further breakdown disconnectedWaitingForVisible (#699)
Replace `disconnectedWaitingForVisible` with `disconnectedWaitingForVisible`, `disconnectedWaitingForVisiblePriorWasInitial` and `disconnectedWaitingForVisiblePriorWasConnectError` so we 
can determine what the prior state of clients not trying to connect because they are hidden was.  This will also
help in further validating the v2 metrics vs the v1 metrics (since most of v1's `never`, are classified in v2 as `disconnectedWaitingForVisible`).