chore: Change the app script name (#706)
Change the app script name to be friendly words followed by the appID encoded as base36. For
example lavish-medieval-meat-lk8oz2zz

Previously it was friendly words followed by the appID. The appID is
base62 encoded matching [a-zA-Z0-9] which includes upper case letters.
This is not compatibly with Cloudflares script names.