perf(reflect-server): don't list all clients every turn (#817)
Tackling this as it showed up at the top of time spent in profiles of reflect.net.
<img width="955" alt="Untitled (1)" src="https://github.com/rocicorp/mono/assets/19158916/0a4d6d78-888e-4063-a32c-e957a4df0a9b">

To enable this had to add getEntries to Storage interface and EntryCache impl.  

We still list all clients if a client needs to be fast forwarded in order to compute lastMutationIDChanges.   To avoid this we need more indexes (a long the lines of faster forward), which is too big of a change to tackle right now.