feat(reflect-cli): reflect dev watch (#906)
esbuild has watch support but it does not work when compiling to a string. Instead we use node:fs watch function.

esbuild outputs the files it is using as inputs (entry point plus dependencies). We setup watchers for these files. When any of them change we remove the watchers and resolve a promise. Then we compile again, getting a new set of files to watch. Rinse and repeat.

The added `watch` function in src/compile.ts yields CompileResult. In dev.ts, we use a loop over this async iterator. When we get a new result, we shut down the dev server and starts a new instance of the dev server.