fix(replicache): improve new client channel message compatability (#928)
Problem
======
Older clients (<= replicache@13.0.1), *asserted* that the messages received on the new client broadcast channel were an array containing exactly one string.

https://github.com/rocicorp/mono/commit/fc30c366f5428583163e00e2c6fb9d9f0f176801 added another string to the message array, which would have caused  <=replicache@13.0.1 clients to throw exceptions in their new client channel message handler, and fail to refresh if the new client had a new client group id.

Solution
=======
Send the old message format, v0, on the old channel name `replicache-new-client-group:${replicacheName}`. 

Send the new message format, v1, on a new channel name `replicache-new-client-group-v1:${replicacheName}`. 

The new message format uses an object instead of an array, and is designed for forward/backward compatibility (going forward).  The new message format can be extended with new optional properties.

Note, this change is not perfectly forward/backwards compatible.  The client no longer listens on the old channel name.  This is because if it did, it would not be able to distinguish v0 messages from  <=replicache@13.0.1  clients (which ideally it would handle) from v0 messages from newer clients (which ideally it would ignore in favor of processing their v1 messages).  

However, it is more compatible than the current code.   <=replicache@13.0.1 will refresh if the >replicache@13.0.1 client has a new client group id (and will not encounter any exceptions due to receiving messages with an unexpected format).  What does not work exactly as desired, is if >replicache@13.0.1  is deployed, but then the deployment is rolled back to <=replicache@13.0.1 , the >replicache@13.0.1 clients will not refresh if new <=replicache@13.0.1 clients have a different clientGroupID.