feat(reflect): Build with code splitting (#965)
Problem
---

The previous build built `reflect-client`, `reflect-server` and
`reflect-shared` in isolation (inside their respective packages). This
meant that the shared code was duplicated in `reflect/client`,
`reflect/server` and `reflect` modules when they get imported from an
app. The code duplication can also lead to `instanceof` not working as
expected because there might be multiple classes that should logically
be the same. I haven't seen this yet but it's a possibility.

Solution
---

Use esbuild to build all three packages in one go and turn on code
splitting. This means that `reflect/client`, `reflect/server` and
`reflect` will all share the common code.

For the .d.ts files we use the old build artifacts because esbuild does
not generate .d.ts files. This could potentially have the same problem
with 2 types not being the same when they should logically be the same.