feat!: Make mutator have var args
BREAKING CHANGE!

Previously, mutators had one optional argument. Now it has variable
number of arguments.

This changes how we treat the `args` property in the local meta commit
which is stored in IDB. We therefore increment the format version.

This changes the `PendingMutation` type which is returned by
`experimentalPendingMutations` API to have an array of arguments.

This changes the push request to have an array of arguments. Therefore
the `pushVersion` is now `2`. The `PushRequest` type is now a union of
`PushRequestV0 | PushRequestV1 | PushRequestV2`.