feat(reflect-cli): Lazy load commands
Running reflect-cli is very slow. It loads 8MB of code before it even
starts. This commit changes the CLI to lazy load commands. This means
that the CLI will only load the code for the command you are running.

There is still a huge chunk of code being loaded because we use the
compat layer for firebase but at least this skips loading workerd if we
are not running reflect dev.

Towards #1087