feat(reflect-client, reflect-server): Implement subscribeToPresence (#1034)
https://www.notion.so/replicache/Simple-Presence-API-dfb13274d7144153a8ae56d29c3e442e

The protocol change is backwards compatible.

Presence is synced via patches.  For a connection the first patch always clears and provides complete current presence state.  Subsequent patches over the same connection are incremental.

Presence patches are at the Poke level in a PokeMessage (a PokeMessage can have many pokes).  Currently the presence patch is always on the first poke of a poke message, but we have the flexibility to change this in the future if we want to more carefully coordinate presence changes with state changes.

On the client presence is stored in memory in a new PresenceManager class. 

Also updates reflect.net to use `subscribeToPresence`.