feat(reflect-server): pass roomID to roomStartHandler (#1133)
Pass roomID to roomStartHandler.

This moves the invocation of roomStartHandler from blockConcurrencyWhile in the BaseRoomDO constructor, to the first fetch.  This was necessary because the roomID is not guaranteed to be known until the first fetch.  Also makes how fetch determines the roomID more robust by having the AuthDO always pass the roomID in a new x-reflect-room-id header.

This also contains a small fix for processRoomStart.  processRoomStart wrote changes at the existing version, rather than the new version.  This is incorrect and can lead to changes not being synced to clients.  Updates processRoomStart to write changes with the new version.