refactor(replicache, reflect)!: Make clientID synchronous, rather than a Promise. (#1197)
`clientID` no longer needs to be a `Promise<string>`, as it is simply a uuid.  `clientID` being a `Promise<string>` creates quite a bit of unnecessary complexity in `Reflect` client's impls and in many of our samples and demos, and more than likely our customers' apps.

BREAKING CHANGE: The type of `Replicache.clientID`  was changed from `Promise<string>` to `string`.  The type of `Reflect.clientID`  was changed from `Promise<ClientID>` to `ClientID`.