feat(reflect-server)!: GC client keys (#1254)
BREAKING CHANGE!

User keys starting with `-/c/<client-id>` are now GC'd after 2 weeks if
the client has not connected to the server in that time.

We keep track of the time a client was last seen by storing a `lastSeen`
property in the client record. This is updated whenever the client
*disconnects* to the server. The reason why it is enough to set this
when the client is disconnected is that we never garbage collect clients
that are currently connected to the server.

When we collect a client, we also collect all the user keys that start with
`-/c/<client-id>`.

The client GC is run every frame. A followup PR will limit this to once
per X minutes/seconds/days. TBD.