fix(replicache): Reject pull/push if closed (#1351)
There was a bug where if you called pull/push on a closed replicache instance, it would never resolve or reject. This fixes that so that the methods reject with an error if the replicache instance is closed.

This got a bit more complicated then I expected because we must not reject promises that has no catch because that leads to log spew. Therefore we keep track of pending push/pull calls so that we only reject if someone called push/pull.