fix: improve errors for invalid room id  (#1344)
Adds a roomID validation check to the Reflect client constructor.  If the roomID is invalid an Error is thrown with message:
`ReflectOptions.roomID must match /^[A-Za-z0-9_/-]+$/.`  

Currently the server returns a `RoomClosed` error if you try to connect to an invalid room id, and a `RoomNotFound` error if you try to tail an invalid room id.  

With this change the server now returns a more useful error:
`Invalid roomID ${roomID} (must match /^[A-Za-z0-9_/-]+$/)`  

This error is returned over the WebSocket in a message and is logged to the console on the client.

Document in ReflectOptions.roomID the character constraints for a roomID.