fix(reflect-server): Deal with pushes of of mutations originating from a deleted client (#1462)
* chore(reflect-server): Get rid of client tombstone

Instead have a `deleted` property on the client record.

If a client with pending mutations gets deleted due to inactivity, and
there are other clients in the client group. When these other clients
do a push we need to update the lastMutationID on the server so that
pull can work correctly and mark the pending mutations as applied.

* fix(reflect-server): Fast forward and pull logic for deleted clients

We need to update the lmid even if the client is deleted. This is
because when a push happens from a different client in the same client
group we need to mark the mutation as processed in pull or we will get
stuck on mutations done on the deleted client.

For mutations created in the deleted client we ignore the effect of the
mutation but we still update the lastMutationID.