chore(replicache): No need to keep #root (#1468)
We used to keep track of the root hash inside the Replicache instance.
The intended use case was to prevent firing changes when the root hash
didn't change.

However when we switched to UUIDs for the hashes we end up generating a
new "hash" even when nothing changes. Therefore we can remove this extra
book keeping.