chore(replicache): Split Replicache Impl (#1631)
Now we have a thin wrapper around the ReplicacheImpl class. The impl
class is what reflect and zero should be using and we will expose the
required APIs on the impl class.

This changes a few data properties to accessors. This should not be a problem unless someone is subclassing Replicache or depend on the enumerable properties of the Replicache instance.