feat(zero-protocol): add push (#1655)
Adds push protocol.

The protocol supports built in CRUD mutations and custom mutations.

Built in CRUD mutations are modeled as a single mutator with the special name "_zero_crud",
that takes an array of CRUD operations. The array of CRUD operations are applied transactionally.

The supported set of CRUD operations are create, set (upsert), update, delete.