chore(zero-cache): Fix tests (#1687)
### Pin vitetest version 1.3.0

This fixes the "No such module" error that was happening when running the tests.

### Add vitest config files for miniflare, node and pg

It turns out that the `@cloudflare/vitest-pool-workers` is not compatible with vitest workspaces. The way `@cloudflare/vitest-pool-workers` works/hacks around a problem is by passing a subclassed array that does a filter `push`.

### Remove workspace config file

Now we have different `vitest.config.*.ts` files instead.

### Change npm scripts

To make more sense now that we have different configurations for different environments.