feat(zero-client): Add `mutate` method to zero (#1727)
The Zero class already has a mutate object. This object has crud methods
for the different registered queries. For example, if you have a query
called issue you can call `zero.mutate.issue.create` to create a new
issue.

This PR changes the mutate object to be a function object. This function
object still has the crud methods for the different registered queries.
But, the function can also be called, in which case it acts as a way to
batch multiple mutations together. For example:

```ts
zero.mutate(async m => {
  await m.issue.create({id: 'a', title: 'Hello', body: 'World'});
  await m.issue.create({id: 'b';, title: 'Goodbye', body: 'World'});
});
```

Under the hood, we only call the `_zero_crud` mutator once for the whole
batch. This is more efficient than calling it once per mutation.