fix(zero-cache): fix error propagation in ViewSyncer invalidation handling (#1788)
The variant of the cursor API that takes a callback did not work as expected.

The cursor call would return a promise that resolved with all of the rows first, and then call the callback on the rows. If one of those callback calls throws an error that error was unhandled.

Switches to use the async iterator variant.

Also updates error handling so that view-syncer clients are disconnected when invalidation error occurs.