refactor(zero-cache,reflect-server): break zero-cache dependency on reflect-server (#1805)
To do this I created a new cf-shared package and factored router and some other utilities from reflect-server into this new package.  This package is separate from shared because its tsconfig needs to include `types: ['@cloudflare/worker-types']`
and it need to not include `lib: ['DOM']`.


I did attempt an approach where I put this in shared and did not use `types: ['@cloudflare/worker-types']` but rather imported the cloudflare types explicitly in the files that used them.  I even updated all of reflect-server and zero-cache to do this explicit importing.  However, I could not get the reflect-server d.ts to build with this approach (which otherwise I liked better).