feat: Add a Float64 encoder that preserves order (#1847)
https://medium.com/@aravindet/order-preserving-encoding-for-floats-cde09c978629

This is a new encoder for Float64 values that preserves lexicographical order. It is
based on the idea of encoding the float as a string that preserves
order. This is useful for indexing float values in a database.

This is represented as base36 since that prevents potential issues when ordering the strings based on the locale/collation.

For example, for all numbers `a` and `b`, the following is always true:

```
a > b === encode(a) > encode(b)
```