preload tweaks (#1933)
Exposes from zero-client a promise that can be awaited to determine when a query preload has completed.

Uses this to better pace the preloading of issue list results and comments.

For comment preloads, pacing ensures only one outstanding comment preload at a time. Other preloads are queued. Issues currently rendered in the UI are scheduled in a high priority preload queue, while those previously in the UI, that didn't get to front of queue while in the UI are scheduled on a low priority preload queue.