explore formulations of iterables driven by callbacks
`join`, `reduce`, `distinct` will need to call PG and be async on the server.

These three operators work by returning iterables.

We can't do `asyncIterable` (because we don't want to break synchrony in the browser) so we'll need a callback driven iterable that can be async on the server and sync in the browser.

Experimenting with formulations of that pattern.

----

Although we may not need this pattern at all!! All values are ready on the PG side.

We can greedily process them rather than lazily process.

And calling next pipeline stage is already asyncable. We greedily process then, when ready, call `newDifference`

Maybe we don't need to make hardly any code changes? We just need to make specific implementations of `join`, `reduce` and `distinct` that are greedy and call PG.