chore(zql): Remove 'matches' concept from left-join operator.
We had this weird 'matches' concept that was really subtle. The
reason we needed it was trying to match the output semantics of
SQL. But we don't need to match the output of SQL because we are
outputting hierarchical data, not tabular data. Taking advantage
of this we can simplify the impl of left-join.

If anyone wants the tabular output they can easily simulate it.