feat(zqlite): updated zqlite table source to support running on the server
- to run ZQL unchanged on the server, `getOrCreateAndMaintainNewHashIndex` needs to be implemented for now
- presumably we want to pass data from `TableTracker` to the source. Add a method to support this.

Follow up:
- The source should update SQLite as it processes diffs. TableSource will be updated to do that while also still allowing batches of writes to flow through.

---

How will SQLite be kept up to date as diffs flow in?

Adds:
1. Add events will be applied to the database by the source immediately
2. The diff will then be sent

Removes:
1. Remove diffs will be sent down pipelines
2. After remove is processed, the remove will be applied to SQLite

Once the `materialiteTransaction` completes, it will `commit` to SQLite in the single-threaded server configuration.

---

Applying adds and removes in this way allows for retrieving history in operators via incremental pull. Why this works is not documented yet.