feat(zql): Implement TS part of select subquery (#2112)
* WIP: Type seems to work so far

* WIP: TypeScript hacking sub queries

* Split EntityQuery into EntityQuery and RelatedEntityQuery

There are now two interfacesL
- EntityQuery
- RelatedEntityQuery. This is what you get from doing `related('x')` and
  it carries the name `x` along with it. It allows the sub query t
  construct the `{x: Return}` type as needed.

The EntityQueryImpl is not exported to ensure that the interface is used
instead.

* Change imports to newEntityQuery

* Basic AST in place.

* Update zero-protocol and zero-cache

To at least make the type check pass

* Add RelationSet

Now the EntityQuery has one more type parameter. It is is set of the
known relations. For now the value of the property is not used.

* Fix type errors in non zql dirs

By putting NoRelations in most places