chore(ivm): Simplify handling of overlay in source.
This approach works by separating the problems. In one generator,
we splice in the overlay without worrying about start position.

Then with the help of a new 'LookaheadIterator' helper class, we
seek to the right start point in the iterator that already has
the overlay in it.

This also allows us to support starting 'before', 'at', and
'after' value that aren't in the dataset. Though I don't think
we need that currently.