feat(zql): Add per-output sorts to Sources.
At first the plan was to require user to create needed indexes
up front.

But on the server this seems like kind of a bummer. Other systems
will make use of good indexes that are available but will muddle
through if needed indexes aren't there.

This is really nice for dx early-on. There's no reason for us not
to do that since we have SQLite.

On the client, all indexes are in-memory. So "muddling through"
would mean creating a sort when needed then discarding it. But
since we are doing IVM we would likely keep the sort around for
life of query. And that's just the same as creating an index on
demand.

This does have downsides if the user is constantly adding and
removing indexes. But that's just the same problem as in a normal
database with not having the right indexes.

Long story short, we'll probably also want to allow user to
force certain indexes to be created. But it's nice to be able
to be able to handle any query on small datasets without config.