refactor(zql): Merge fetch and hydrate.  (#2177)
The distinction between fetch and hydrate was complicating the take implementation.  In particular it was difficult for take to determine if it should use hydrate or fetch when backfilling to limit.  The next row might have been one previously hydrate that was pushed out of its window by an add, in which case backfilling should use fetch, but if the next row had not been previously hydrated, hydrate needed to be used.  Additionally hydrate would need to take a start argument.  

We've decided to remove the distinction between fetch and hydrate and require individual operators to keep enough state to correctly handle fetching the same row multiple times.   This required join to store a bit more state to correctly handle cleanup.  Before join stored 
`['hydrate-count', parentKey] => count`
 and now stores, 
 `['pKeySet', parentKey, ...parentPrimaryKey] => true`
 
Before it would clean up the child for `parentKey` when a removes causes the corresponding `hydrate-count` count to become `0`.  Now it cleans up the child for `parentKey` when a remove cause there to be no keys with the prefix `['pKeySet', parentKey]`.

As part of this change dehydrate has been renamed cleanup.  In a follow up change fetch will be renamed pull.