Simplify Operator interfaces more.
Having the methods parameterized by the caller was a bad idea.

Instead, have operators that have more than one input or output
handle tracking who they are talking to internally.

Currently, this is Source (more than one Output) and Join (more
than one Input). But in the future we will have this same
problem with supporting OR, since the pipeline will have to
fork there too.

The way this is done abstractly is that when you connect() to a
Source, sources gives you a specialized implementation of Input
that knows how to deal with the fact that Source gets called by
more than one object.

And similarly when Join calls setOutput() on its parent and
child Inputs, it provides a specialized Output that knows how to
deal with the fact that pushes can come from both.

In Source this is complicated by the fact that TypeScript lacks
private classes so it's somewhat hard not to leak the details,
but we accomplish it with closure shenanigans.

This all has the interesting side-effect that not all "Operators"
are now both Input and Output. A Join is an Input, but not an
Output. We may want to remove the Operator interface completely
though and just let simple operators implement both Input and
Output.