chore(zero): Remove zero-cache and zero-client again.
I do not understand why this is necessary. For some reason
including these causes npm to try and install these packages
when zero is added to an existing project.

However it doesn't do that with Replicache. I do not understand
the difference.

Luckily with create-canary.js this is less of a problem since we
always build from a fesh directory.

It should still be understood though. I just have to get this
release done.