chore: Use relative paths everywhere (#2590)
This changes to use relative paths everywhere.

The only non relative paths remaining are `@rocicorp/zero` in `apps/zbugs`. These should remain because this should use the "built" version to use the public API.

This removes the "paths" in all of our tsconfigs. It also removes the vite-tsconfig-path plugin.

This was done using a script (transform.ts in the first commit) and then this script was run over the different packages.

This also adds a eslint rule that prevents using 'packagename/...' and 'packagename'.