fix(zero-client): CRUD compound primary key (#2629)
* fix(zero-client): CRUD compound primary key

We now support compound primary keys in CRUD operations.

These primary key can have arbitrary name and have the type
`number | string | boolean`. The type is not yet enforced in the schema
type but it is enforced at runtime.

The key that we use in Replicache is:
- If there is 1 primary key (e.g. `['id']`) we use the value of the
  primary key.
  `e/<tableName>/<valueOfPrimaryKey>`.
- If there are more than 1 primary key (compound primary key) we use a
  hash of the primary key values. We do not include the names of the
  primary keys in the key. This is because the primary key names are
  fixed by the schema and the table name already needs to be unique.
  `e/<tableName>/<hash>`.

For the hash we JSON.stringify the primary key values and then hash
them using `h64` with a reverse. This is the same technique that we use
in zero-cache to generate the row key.

* feat(zero): Tighten CRUD param types

Now we enforce that the value/deleteID matches the primary key type