fix(zql): edit with nested joins (#2686)
When we have nested joins we were not preserving relationships correctly.

We need keep the existing relations so converting an `edit` to a `remove` + `add` is not correct because the remove would remove the existing relations. Instead we keep the `edit` which changes the row without changing its relations. Then if the `edit` caused the join key value to change we fetch the child stream and creates multiple `child`+`remove` changes which removes all the rows from the relation followed by fetching the child stream and creating `child`+`add`.

Fixes https://bugs.rocicorp.dev/issue/k_h0Dy_6_6yHFDWFxNke2