chore(replicache): Switch to vitest (#2710)
This involved a few things:

The `package.json` scripts now have `test-chromium`, `test-firefox` and `test-webkit`. `test` just calls these in sequence. It is less efficient than running them concurrently but overall it is still faster.

`worker-test.ts` (the file loaded in a worker) was pulling in `vitest` which was not allowed. I removed the deps in there.

All the tests files now import `expect`, `test`, `describe` etc