fix(zbugs): fix read state flicker (#2814)
Fixes https://bugs.rocicorp.dev/issue/3078

There would sometimes be a flash of the unread indicator on read issues in list page.

This was hybrid query + consistency issue.

The preload query did not include the 'viewState' relation. So before the list query was got, the viewState could be undefined (as it was not covered by preload) which was interpreted as unread.