WIP: Ideas for materialize with a factory (#2827)
This changes materialize to allow passing in a function. The function is responsible for creating the view.

The intent is to allow Solid to create a different view for materialize. It would be used something like:

```js
const solidView = q.materialize(solidMaterialViewFactory);
```

The API is still pretty much work in progress.