fix: Remove top level await from xxhash
Instead, start initiating the wasm as the module is loaded. Then expose
a promise that resolves when the wasm is ready to be used.

The functions will assert that the XXHash API has been initialized
before usage.

Both zero-cache and zero-client have been updated to wait for the wasm
ready promise before depending on the xxhash functions.

It is possible that there are other code paths that leads to these
functions so keep your eyes open.