feat(zero): move schema into its own package
The schema needs to be available on the client and server. Exporting the schema from `zero-client` or `zero-cache` doesn't fit well in this model. Created a new package `zero-schema` that only depends on `zero-protocol` and `zql`.

The dependency on `zql` is needed since auth will be in the schema and expressed as queries.