fix(zbugs): add permissions for emoji and user prefs (#3184)
An emoji can only be updated by its creator. It can only be seen if the
user can see the issue it is attached to.

A user pref can only be updated by the user it belongs to.

Also add "can see" to a few other insert, update and delete permissions.