fix(zbugs): Emoji toast (#3224)
The toast was not working due to it being triggered by the emoji panel.
However, the emoji panel was lazily rendered, so the toast was only
working if a new emoji reaction was added to one of the comments that
were currently visible.

Now, we subscribe to all emojis attached to the current issue and all
the issues in the comments of the current issue (using `IN`). This way,
the toast will work even if the emoji panel is not rendered.