chore(zbugs): Improve rendering perfomance (#3238)
Change comment list to only overscan by 2 (was 5). Rendering comments is expensive due to markdown. We should really deal with the markdown once on the server instead of doing it on the client on every react render.

We were also over rending all over the place... By adding memo/useMemo/useCallback in a few places we can improver rendering further. These are just a few things I ran into when looking at the over rendering of rows in the lists and it is not an exhaustive fix.