fix(zbugs): Do not render all emoji toasts (#3239)
The old strategy of using the time between the first render and the time
that we got the emoji data does not work during initial sync because the
time it takes to get the data can be large.

Instead, we use the newly added `ResultType` from the `QueryResultDetails`
to determine if this change is part of the initial sync or not. Once we
get a `'complete'` result type, we start treating changes to the emoji
array as new emojis.

This changes to use stacked toasts which look nicer when there are
multiple toasts on the screen.

This change also improves the auto close behavior of the toast. toastify
has support for auto close but it is not working correctly. Instead we
turn off the auto close and manage it using `setTimeout` and listening
to focus/blur events.