fix(zql): Remove assert that start must match constraint (#3243)
This assert has the comment:
```
        // There's no problem supporting startAt outside of constraints, but I
        // don't think we have a use case for this – if we see it, it's probably
        // a bug.
```
This was probably true when there was just filter, take and join.

However, now there is a case where this happens, specifically with skip.  

Scenario:

Query has a related and a startAt (i.e. Join and Skip operators).

1. join receives a push from child input
2. join fetches from parent input with `constraint: { parentKey: parentValue }`.
3. skip fetches from input with `constraint: { parentKey: parentValue }, startAt: bound`.

Skips bound is whatever row the user passed to query.startAt, and it may not match constraint.

This assert was only ever done in MemorySource, so TableSource did not have to be updated.