feat(zbugs): Show emoji tooltip on pill when added (#3250)
When a new emoji reaction comes in we show a tooltip on the emoji pill that represents that reaction. This tooltip will show the user who added the reaction.

The logic for this is pretty complex.
1. We keep track of recent emojis
2. We show the tooltip if the emoji is in the recent emojis
3. We hide the tooltip after a timeout. But we only start the timeout:
    1. When the pill is in view
    2. The document has focus.

If there are multiple emojis matching the pill we show them one at a time.  If the user clicks on the toast we reorganize the order to ensure that the clicked emoji is shown first.