feat(zero-cache)!: more detailed jwt options
- JWTs can be verified either by public or private key.
- JWT verification keys can be discovered from a remote endpoint
- JWTs can be encrypted in addition to being signed

This adds options to:

1. Use a public or secret key for verification
2. Use a JWK url to discover keys for verification
3. Specify a secret key to use for decryption

From what I can tell, it is not possible to discover the decryption key from a JWK remote.
This makes sense to me. The keyset service doesn't know
who is getting keys from it so it can only provide public keys. To decrypt
a JWT with a public key would require that JWT to be encrypted with our own
private key. Something the discovery service won't have.