fix(zbugs): I give up on the scrolling. (#3414)
Changed the overscan to 3 to make it slightly less likely to hit it.

Also, search more comments before giving up.

The reason it does not work is that the view scrolls before we can
compute the anchor so we end up using the new anchor when the rows are
removed (count of removed rows > overscan).